Added base interfaces for messaging and a post office for sending messages
Added an IListener interface which is able to recieve messages. Added an IMessenger interface which is able to send messages to IListeners through IListeners subscribing to an IMessenger. Added a PostOffice class which is a singleton and is used to hold all unique IMessengers and to register IListener to specific MessageTypes.