Added a "system" which is a part of the engine which can communicate with other systems via the post office
Added a system: SceneBuilder. This is a part of the engine and it is decoupled from all other parts of the engine. All systems are initalised and managed by the SystemManager class which lives inside the Engine instance. The Engine instance requests a new scene be built via the PostOffice and the SceneBuilder returns it again via the PostOffice. Scene built is currently hardcoded and empty.