Refactored some of the PostOffice code. Removed usage of PostOfficeMessenger.
The PostOffice has been simplified slightly for now. It's usage currently requires a class to inherit PostOfficeListener and to subscribe to a MessageType (defined in Consts.h). When any other class sends a message of that type to the PostOffice singleton, it will get sent to the class subscribed to that type. Example usage of this is shown between the Engine and SceneBuilder classes.