Removed IScene from the Engine class. Added a SceneManager system
Added a new system: SceneManager. The job of this system is to update and draw the current scene, as well as recieve a new build of a scene from the SceneBuilder system. The reason for this change is to seperate scene and engine information, achieved by removing IScene references in the Engine class.