Implemented scene building from file
The engine now creates scenes from file. Originally scenes were going to be stored in JSON but I decided XML was easier to read as raw data and easier to implement. Added a new ComponentBuilder class whose job is to simply parse XML component nodes and build the correct comonent from it and add it to the GameObject. There is only one scene so far, stored in the resources folder. This scene is selected via a hard-coded string in the engine class.