SystemMessageDispatcher singleton moved into a seperate class to make unit testing easier
Reason for this is because it is super hard to unit test a singleton, a static instance. This way, the class with the messaging functionality can be instantiated without any of the singleton nonsense. Added a test class to test the messaging system. Also cleaned up #includes in various files.