Removed the singleton message dispatcher all together
Instead of going for the singleton approach to a centralised messaging system, I have changed it to a dependency injection approach. The SystemManager class now holds an instance of the SystemMessageDispatcher. All systems now hold a pointer to this instance and it is injected through the systems constructor. To listen for messages the system simply needs to call SubscribeToMessageType(). To send messages, the system needs to additionally subscribe to SystemMessageMessenger, and from there can call SendMessageToDispatcher. I decided to seperate sending messages from ISystem because some systems do not need send messages. This may be changed in the future.