Changed the way that IDrawableComponents should be drawn
Since removing the globally accessible singleton, there was no way for IDrawableComponents to contact the Graphics system. To fix this, the SceneManager class now provides all drawable GameObjects with a DrawSceneMessage reference that they can add drawing information to. After every GameObject has added to the message, it is sent to the Graphics system where it is used to draw the scene.