Fixed issue with logger singleton.
Previously every time you sent a message to be logged, it would create a new spdlogger instance which was not how the developers designed it and it would crash. Changed it so that there is only one instance of spdlogger. Moved the LogSeverity enum out of Consts.h so that to log something you only have to include Logger.h and not Consts.h