Moved drawing functionality from Scene to Graphics
Multiple changes related to drawing. The DrawSceneMessage no longer holds two vectors of Sprite and Text data. Instead it simply contains a pointer to the current scene which the IGraphics instance extracts IDrawableComponents from.

The IDrawableComponent has been changed to simply contain data rather than have an overridable function. The idea behind this is that when various drawing components implement it (such as SpriteRendererComponent) additional information will be added such as position, rotation etc. In IGraphics when this component is recieved it will need to be dynamically casted into the appropiate component type and the information extracted from it and used to draw.

Other changes include renaming IScene to Scene. This is because I don't currently think there needs to be multiple scene implementations.