Added an enum to represent the drawing component type. Moved some functions out of the IGraphics interface
Instead of the IGraphics interface knowing about all the drawable component types and having the job of dynamically casting them, the responsbility now falls to the dervied classes. This also means the interface will never have to change.

Added an enum to IDrawableComponent which makes it easier and quicker to cast it to the correct drawing component.