Removed scene initalisation from Engine system
Instead of the RequestBuildSceneMessage being sent from Engine in the SystemInitalised() function, it is now sent from the SystemManager after everything has been initalised. The reason for this is because the Engine system should be the same with or without the Editor being run too. However with the Editor we don't want the engine to start of in a scene, rather we want the user to pick a scene. In the SystemManager class there are two constructors, one for the Standalone version and one for the Editor version which is a much better place to send the message.