Various changes to accomodate the external editor
Changed SystemManager back to holding a map of Systems rather than a vector.

Created a Editor System within the engine that will be used to talk to the rest of the engine through the messaging system. The engine will still need an interface of static extern functions that the editor can actually call.