Changed Engine from being a System to taking over the role of SystemManager
Previously there was an Engine class that was an ISystem. This would recieve messages from SystemManager and would simply update other systems in a specific order. However this meant that two classes did exactly the same thing: update other systems. This didn't make sense so I removed the old Engine System and gave SystemManager it's responsbilities. This now includes initalising and storing all systems and now includes updating them in a specific order. Also renamed the class to Engine.

Removed a system test about the Engine System as there is no longer an engine system.