Fixed the issue with strings not being saved into XML and also parsing/writing GUIDS
Solved a long-standing issue of parsing a GUID to and from a string. Previously all attempts would corrupt the stack and the engine would simply generate a new GUID each time and ignore the guid set in XML. However now it successfully reads the GUID in from the XML file and sets it to the GameObject.

Saving a scene now works as intended, including saving the GameObjects GUID.