Editor can now play and stop scenes
When run with the editor, the central update loop is now started on engine initilisation. The scene update and physics update are not run until play has been pressed in the editor however the drawing is. When play mode is stopped in the editor the current scene is rebuilt from the XML file.

In the future, when the editor can change values in GameObject components, the scene will need to be saved to XML before play mode can be entered.

Temporarily removed font initilsation in TestGraphics class while I work out a way to best to share resources between standalone and editor mode.