WIP on fetching component information from the hierarchy
IComponent now has a pure virtual function: ExtractComponent(). This returns a map with string keys and values containing information about variables within each component. Every component must override it.

A specific GameObject in a scene can be fetched using the first part of the GUID. This isn't future proof as the data1 part alone of a GUID has the potential to be duplicated.