Scene building/saving improvements
In SceneBuilder, the order of scene creation has changed. Firstly all GameObject's get initalised. Once that is all done, the loop is repeated and all GameObject's have their components built. And finally the loop is repeated again and all GameObject's are checked to see if they have parents, and if they do the parent is set.

You can now save GameObject's parents in the scene file. This is done by specifying the parent's GUID in the "parent" attribute on the GameObject node.

The foundations for mapping XML attributes to GameObject/component pointers has been setup. To fetch a GameObject pointer or component pointer in a component build function (found in ComponentBuilder.cpp) you simply call _scene->GetGameObject(GUID), with GUID being the GUID of the GameObject. This will return a pointer to the GameObject and from here you can call GetComponent to get a pointer to a specific component.

All components now have the GUID of the GameObject they belong to. This can be used in the ComponetnSaver to represent a pointer to this current GameObject.

Logger now clears all entries on application start.