SpriteRenderercomponent and TextRendererComponent modified + "expansion" of ComponentFactory
SpriteRendererComponent and TextRendererComponent now contain a pointer TransformComponent (which has the attributes position, scale and rotation). They have bee modified because a sprite (and a text) are both transform components, hence it had no meaning to add them a position(etc) while it had to be represented by the Transformcomponent that would be modified in the other parts of the program.
Furthermore, some modifications have been added to Component factory : added methods to create sprite & text components. We will have to delete the transformComponent creation in the factory since this component would either be a text or a sprite. What we could also do would be to pass in a transform parameter directly in those constructor instead of taking its attributes in the constructor (_localPosition, _localRotation and _localScale).