ComponentBuilder modifications
Some modifications have been done in component builder, accordingly to the previous modifications done in ComponentFactory. However, we might need to change the way we define elements to draw in the scenes in the xml files : first, we would need to specifiy wether these are text or sprite (so that we can send them to the corresponding methods when we build the components). Also, we would need to specify the color & text for the text, the path for sprite and the rect and offset for both.