Multiple changes related to ComponentFactory and scene building
Added a step in scene building: dependency fetching. The purpose of this step is for components to fetch pointers to anything that component needs, most likely other components belonging to the same or to other GameObjects. The reason this has change has been made is because previously the ComponentBuilder was attempting to fetch pointers during the step where components were being built, however sometimes the component the pointer was pointing to hadn't been instantiated. This way all components on all GameObjects will be instantiated, then pointers to required components can be fetched by other components.

Added a new component interface: IStartableComponent. If a component chooses to implement this interface, a Start function will be called when play mode is entered.