Camera class changed - Some issues to fix
Camera changed to a 2D game camera instead of a 3D game camera.  Should add some more fonctions, like center to a game entity for example but might depend on the game we want to implement... Also, some issues appeared after last changes with TextRendererComponent and Sprite RendererComponent... Been trying to fix it this afternnoon - still on it so far (most probably a messaging problem to the graphics class)
Also ; been changing Component Builder to retrieve different values for Text and Sprite renderer components.