TextureManager v0.1
Added class Texture manager but so far  not used yet. The textures are managed with _textures map in DxGraphics. Loads a texture by knowing its path with LoadTexture method and retrieves it with GetTexture method. Still have to deal with the operations on the sprite for now  (rotation, scale, etc.)