Animation0.3
Animation class updated and added to DxGraphics. Now, DxGraphics can load the list of the animations names from xml file /Resources/Animations.xml, then it will load the corresponding animations according to the saved names. However, we should change the Sprite renderer component to add the animation  we want it to do. For this, we could add 2 elements in it : objectName + action. If so, we would have to save the animation names as so (ex : Player2Walk).