Animation 0.4
SpriteRendererComponent changed. It now contains Name and Action, which corresponds to the animation we want it to do. Ex : if we want the "PlayerGreen" to walk, name will be "PlayerGreen" and the animation parameter will be "Walk". The corrsponding animation will be done (it has to be called PlayerGreenWalk.xml" too.