Setup project to build 'Games' alongside the engine
Setup ComponentFactory and ComponentBuilder to use components that are not 'core' components, rather belong to a specific game such as PlayerController. Started work on AlexGame with a new XML scene file and several GameObjects.

Fixed a PhysicsSystem bug where collisions weren't being registered. This was caused by the ObjectGrid not being setup with a big enough size and step to encompass all objects.