Configured the engine to take in a string that points it to the location of the resources folder
Since there are 3 different ways to launch the engine, each with a different project and folder, there are issues with fetching resources from file, simply because the path specified for when the engine is run as a standalone is not the same as when it is run with unit tests. To fix this, a path to the resources folder must be provided when creating an instance of the engine.

For standalone, this is specified in main.cpp and is the same as it used to be: "..//Resources". For tests, the current cpp class is queried to find its directory location and then the string concatenated to get to the resources folder. For editor builds, where the DLL of the engine could be in any random place, the editor passes the location of the resources folder to the engine, which is easy to determine in C#.

The path to the resources folder is now globally accessible, accessed through a singleton instance called GlobalVariables.

I have also moved the input maps to the resources folder. For some reason the audio system was no longer being initalised so I fixed that too.