Improvements to player-platform interactions
Collisions between platform and player are now accurate thanks to a small tweak in SpriteRendererComponent. The tweak changes GetPosition() to return the centre point of a sprite no instead of the scaled/rotated top left point. Also changed all IRenderableComponent references to TransformComponent that attempted to modify/fetch local values to point to world values instead.

Added a new MainWindow message that gets fired out whenever the keyboard gains focus on the application. This is meant to complement the lost focus that current exists. In future it may be useful to have certain systems listen for these messages and then stop updating until focus comes back.

Further work on AlexGame and general fixes.