Added message notifications to GameObject collisions
When two ColliderComponents collide, their GameObject's will recieve a CollisionMessage containing a shared_ptr to and the tag of the collided GameObject. This functionality is currently being used in AlexControllerComponent, where a check is in place to see if the Player collides with a GameObject with the tag of "Platform" (and then in the future it will bounce).

Added DoodleJump related sprites.