Implemented jumping mechanic for player
Player now jumps when landing on platform. This is done through explicitly setting the player's rigidbody velocity to a set value via a message. Also adding a check that makes sure the jump doesn't trigger too often which could potentially happen if the player collides with the side of the platform collider.