Replaced individual platforms with a PlatformManager component
The PlatformManager component now owns all of the platform GameObjects in an object pool. In future it will be used to spawn/move platforms above the player when the player reaches a certain height, creating the endless gameplay.

Modified the IGraphics DrawScene case statement to not include objects for drawing that are currently inactive.

Moved the StringToGUID function from SceneBuilder to ComponentBuilder so it can access it too. Also made it static.