Added platform movement and added a GameManager component
The platforms now move in a manner similar to the actual Doodle Jump game, in that when a player gets to a certain height, they move up. The next step will be spawning more platforms when the player gets to that height.

Added a GameManager component which will be the component that has control over all aspects of the game.