Fixed issue where collider positions weren't updating
This issue occured because the parent transform was being modified, which in turn affect child transforms world positions. However because the parent transform was the one being directly modified and not the children, the child transforms HasChanged flag (which is set to true when a transform value changes) wasn't getting flagged, only the parent's was. To fix this the CheckChanged() function in TransformComponent now checks if the the transform has a parent and if so checks if their transform has changed, since that will change their transform anyway.

Also changed the size of the ObjectGrid.