Added random platform generation to AlexGame
The platform positions are now randomly generated using the <random> library. The random number for the X position is between two horizontal boundaries that will get bigger over time. The random number for the Y position is between a const min height and the current max height which will get bigger over time. This produces platforms that are fairly easy to navigate around but will get harder and further apart as the game progresses.

Fixed the tests project by including newly created files.