Player height is now limited to halfway and added game progression
The player height is now limited to the middle of the screen. If the player climbs over that then the player's height is artificially limited there, however the player's velocity remains unchanged. Using the players velocity the platforms around the player can move upward at the same speed as the player was, giving the illusion that the player is still moving upwards.

Added game progression to the game which is determined by a number of factors. Firstly a 'terminal' score has been defined, which is basically the score where the game should be at it's hardest. Then the two progressive values get interpolated between their minimum and maximum by the amount that the current score has reached the terminal score.