Fixed various bugs, broken unit test and added side warping to AlexGame
The unit test, CoreEngineUpdateLoop, has been failing for a while because it was stuck in the main engine update loop. Turns out this was simply because '_lastTime' hadn't been initalised in the engine.

When running a scene from the editor there were some weird physics bugs where collisions were happening with objects that weren't there. This was because objects from the previous scene run hadn't been cleared from the vectors. These vectors are now cleared on PlayStop. Same with the audio system.

There was a bug with deleting components from a GameObject, specificially rendering components. Attempting to delete these components as IComponents threw an error. To fix this, the first class these components inherited from was set to IComponent instead of IRenderableComponent.

Fixed a math error in sprite renderer component where the Y position was being incorrectly calculated.

Made it so rigidbodies are now fully uneffected by the physics system. Previously they were only partly effected.

In AlexGame made it so if the player goes off a side they warp to the other side.